Moved all the configurations into the configs/ subfolder to separate the test scripts from the configurations. Updated __init__.py and other references accordingly.
