+++ python/lsst/ts/IntegrationTests/__init__.py
@@ -27,18 +27,10 @@ except ImportError:
 from .base_script import *
 from .testutils import *
 from .yaml_test_strings import *
-from .shutdown_configs import *
-from .take_image_latiss_configs import *
-from .track_target_configs import *
 from .auxtel_prepare_for_onsky import *
 from .auxtel_visit import *
 from .auxtel_shutdown import *
 from .auxtel_track_target import *
-from .auxtel_state_transition_configs import *
-from .maintel_state_transition_configs import *
-from .obssys2_state_transition_configs import *
-from .eas_state_transition_configs import *
-from .enabled_offline_state_transition_configs import *
 from .auxtel_offline_standby import *
 from .auxtel_standby_disabled import *
 from .auxtel_disabled_enabled import *
