+++ python/lsst/ts/IntegrationTests/auxtel_disabled_enabled.py
@@ -21,7 +21,7 @@
 __all__ = ["AuxTelDisabledEnabled"]
 
 from lsst.ts.IntegrationTests import BaseScript
-from .config_registry import registry
+from .configs.config_registry import registry
 
 
 class AuxTelDisabledEnabled(BaseScript):
