+++ python/lsst/ts/IntegrationTests/auxtel_shutdown.py
@@ -21,7 +21,7 @@
 __all__ = ["AuxTelShutdown"]
 
 from lsst.ts.IntegrationTests import BaseScript
-from .config_registry import registry
+from .configs.config_registry import registry
 
 
 class AuxTelShutdown(BaseScript):
