+++ python/lsst/ts/IntegrationTests/auxtel_standby_disabled.py
@@ -21,7 +21,7 @@
 __all__ = ["AuxTelStandbyDisabled"]
 
 from lsst.ts.IntegrationTests import BaseScript
-from .config_registry import registry
+from .configs.config_registry import registry
 
 
 class AuxTelStandbyDisabled(BaseScript):
