+++ python/lsst/ts/IntegrationTests/auxtel_visit.py
@@ -21,7 +21,7 @@
 __all__ = ["AuxTelVisit"]
 
 from lsst.ts.IntegrationTests import BaseScript
-from .config_registry import registry
+from .configs.config_registry import registry
 
 
 class AuxTelVisit(BaseScript):
