+++ python/lsst/ts/IntegrationTests/configs/auxtel_state_transition_configs.py
