+++ python/lsst/ts/IntegrationTests/configs/config_registry.py
