+++ python/lsst/ts/IntegrationTests/configs/eas_state_transition_configs.py
