+++ python/lsst/ts/IntegrationTests/configs/enabled_offline_state_transition_configs.py
