+++ python/lsst/ts/IntegrationTests/configs/maintel_state_transition_configs.py
