+++ python/lsst/ts/IntegrationTests/configs/obssys2_state_transition_configs.py
