+++ python/lsst/ts/IntegrationTests/configs/shutdown_configs.py
