+++ python/lsst/ts/IntegrationTests/configs/take_image_latiss_configs.py
