+++ python/lsst/ts/IntegrationTests/enabled_offline.py
@@ -21,7 +21,7 @@
 __all__ = ["EnabledOffline"]
 
 from lsst.ts.IntegrationTests import BaseScript
-from .config_registry import registry
+from .configs.config_registry import registry
 
 
 class EnabledOffline(BaseScript):
