+++ python/lsst/ts/IntegrationTests/maintel_disabled_enabled.py
@@ -21,7 +21,7 @@
 __all__ = ["MainTelDisabledEnabled"]
 
 from lsst.ts.IntegrationTests import BaseScript
-from .config_registry import registry
+from .configs.config_registry import registry
 
 
 class MainTelDisabledEnabled(BaseScript):
