+++ python/lsst/ts/IntegrationTests/maintel_offline_standby.py
@@ -21,7 +21,7 @@
 __all__ = ["MainTelOfflineStandby"]
 
 from lsst.ts.IntegrationTests import BaseScript
-from .config_registry import registry
+from .configs.config_registry import registry
 
 
 class MainTelOfflineStandby(BaseScript):
