+++ python/lsst/ts/IntegrationTests/maintel_standby_disabled.py
@@ -21,7 +21,7 @@
 __all__ = ["MainTelStandbyDisabled"]
 
 from lsst.ts.IntegrationTests import BaseScript
-from .config_registry import registry
+from .configs.config_registry import registry
 
 
 class MainTelStandbyDisabled(BaseScript):
