+++ python/lsst/ts/IntegrationTests/obssys2_disabled_enabled.py
@@ -21,7 +21,7 @@
 __all__ = ["ObsSys2DisabledEnabled"]
 
 from lsst.ts.IntegrationTests import BaseScript
-from .config_registry import registry
+from .configs.config_registry import registry
 
 
 class ObsSys2DisabledEnabled(BaseScript):
